Add truss edit dialog
