+++ docs/release-notes-draft.md
@@ -6,7 +6,7 @@ Changes since **v1.4.0**.
 
 ## New features
 
-- Added an Edit Truss dialog from double-clicking the Trusses table, including MVR instance fields, GDTF truss type fields, and automatic Perastage GDTF creation when type metadata is edited on a model-only truss.
+- Added an Edit Truss dialog from double-clicking the Trusses table, including MVR instance fields, GDTF metadata, a reusable 3D preview, GDTF truss type fields, and automatic Perastage GDTF creation when type metadata is edited on a model-only truss.
 - Added MVR-xchange remote file requests with a larger selectable advertised-MVR list, Console-styled transfer log, corrected station-name alignment, and the standard import choice to open as a new project or merge into the current project.
 - Added the first MVR-xchange TCP Mode publisher so Perastage can manually publish the current scene as an MVR revision for compatible clients.
